requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
Werkzeug==3.0.2
pandas==2.2.2
openpyxl==3.1.2
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
cryptography==43.0.1
python-dotenv==1.0.1